NODE: feat(cryptoCallbacks): pass crypto errors to libmongocrypt
Passes crypto errors that occur on the js side back to
libmongocrypt. Requires every cryptoCallback to return a number
on success, or an error on failure.

Also required an update to the state machine logic to account for
an error occurring during mongocrypt_ctx_finalize.

Note: errors in cryptoCallbacks sha256 and hmacSha256 will
cause a segfault until CDRIVER-3282 is fixed

Fixes NODE-2082